refactor(test): rename tests directories for better clarity
