+++ packages/runner/test/integration/fixtures/instant-generator.js
