+++ packages/runner/test/integration/fixtures/sigterm-ignoring-generator.mjs
